moved IPeriodeEncodageNotesTranslator to common_domain
