+++ ddd/logic/encodage_des_notes/common_domain/service/i_periode_encodage_notes.py
