+++ ddd/logic/encodage_des_notes/encodage/domain/service/feuille_de_notes_gestionnaire.py
@@ -30,7 +30,7 @@ from ddd.logic.encodage_des_notes.encodage.domain.service.i_feuille_de_notes_ens
 from ddd.logic.encodage_des_notes.encodage.dtos import FeuilleDeNotesGestionnaireDTO, NoteEtudiantDTO
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_attribution_enseignant import \
     IAttributionEnseignantTranslator
-from ddd.logic.encodage_des_notes.soumission.domain.service.i_periode_encodage_notes import \
+from ddd.logic.encodage_des_notes.common_domain.service.i_periode_encodage_notes import \
     IPeriodeEncodageNotesTranslator
 from osis_common.ddd import interface
