+++ ddd/logic/encodage_des_notes/soumission/domain/service/feuille_de_notes_enseignant.py
@@ -25,14 +25,12 @@
 ##############################################################################
 from typing import List, Dict
 
-from ddd.logic.encodage_des_notes.soumission.builder.responsable_de_notes_identity_builder import \
-    ResponsableDeNotesIdentityBuilder
 from ddd.logic.encodage_des_notes.soumission.domain.model._note_etudiant import Noma
 from ddd.logic.encodage_des_notes.soumission.domain.model.feuille_de_notes import FeuilleDeNotes
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_attribution_enseignant import \
     IAttributionEnseignantTranslator
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_inscription_examen import IInscriptionExamenTranslator
-from ddd.logic.encodage_des_notes.soumission.domain.service.i_periode_encodage_notes import \
+from ddd.logic.encodage_des_notes.common_domain.service.i_periode_encodage_notes import \
     IPeriodeEncodageNotesTranslator
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_signaletique_etudiant import \
     ISignaletiqueEtudiantTranslator
